+++ Assets/scripts/utilities/version-utility.js
@@ -1,7 +1,7 @@
 import { MOD_DATA, returnModData } from "../modding/data.js";
 
 export const VERSION_DATA = {
-  version: "0.3.0",
+  version: "0.3.1",
   version_label: "",
   dev_label: "(PROTOTYPE)",
   missing_type: "(NULL VERSION TYPE)",
@@ -22,8 +22,7 @@ export function returnVersion(type = "version") {
   var devLabel_full = VERSION_DATA.dev_mode ? nlc + devLabel : "";
 
   var version = (MOD_DATA.enabled) ? MOD_DATA.version : VERSION_DATA.version;
-  var version_nicodrag = (MOD_DATA.enabled) ? nlc + MOD_DATA.nicodrag_version : nlc + VERSION_DATA.version_nicodrag;
-  var versionLabel = (MOD_DATA.enabled)? '' : nlc + VERSION_DATA.version_label;
+  var versionLabel = (MOD_DATA.enabled)? '' : (VERSION_DATA.version_label.length > 0) ? nlc : '' + VERSION_DATA.version_label;
   
   switch (type.toLowerCase()) {
     case "version":