+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Page titles now have "DragginSite - " added to it if it isn't the home page
 - Credits Page
   - New JS Files: `utilities/creditsUtil.js`, `credits.js`
+- Credits Page navigation Text
 ### Removed
 - " - home" from Home Page Title